refactor: Update imports to use relative import paths in containers, main, and ssh modules
