+++ src/containers.py
@@ -1,9 +1,9 @@
 import logging
 
-from rich import print
+from rich import print  # pylint: disable=redefined-builtin
 from rich.table import Table
 
-from src.constants import console
+from constants import console
 
 logger = logging.getLogger("rich")
